// src/app/angular-client2/model/maintenanceDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MaintenanceDto { 
    id?: number;
    cost?: number;
    task?: string;
    date?: Date;
    vehicleId?: number;
    vehicleModel?: string;
    status?: string;
}